data-manager/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
gevent>=23.9.0
gevent-websocket>=0.10.1
gunicorn>=21.2.0
psutil>=5.9.0
requests>=2.31.0
yfinance>=0.2.0
pandas>=1.5.0
numpy>=1.23.0
pytz>=2023.3
python-dotenv>=1.0.0
